:hammer: rename types
