+++ index.js
@@ -1,2 +1,2 @@
-!function(t,n){"object"==typeof exports&&"undefined"!=typeof module?n(exports,require("lodash/at"),require("lodash/omit"),require("lodash/uniq"),require("lodash/without"),require("lodash/keyBy")):"function"==typeof define&&define.amd?define(["exports","lodash/at","lodash/omit","lodash/uniq","lodash/without","lodash/keyBy"],n):n(t["trampss-redux-factory"]=t["trampss-redux-factory"]||{},t._at,t._omit,t._uniq,t._without,t._keyBy)}(this,function(t,n,r,e,u,i){"use strict";n=n&&"default"in n?n.default:n,r=r&&"default"in r?r.default:r,e=e&&"default"in e?e.default:e,u=u&&"default"in u?u.default:u,i=i&&"default"in i?i.default:i;var a=function(t){return"@trampss/"+t.toUpperCase()},o=function(t){return a("SET_"+t)},c=function(t){return a("RESET_"+t)},f=function(t){return a("ADD_"+t)},d=function(t){return a("UPDATE_"+t)},s=function(t){return a("REMOVE_"+t)},l=Object.freeze({SET:o,set:function(t){return function(n){return{type:o(t),payload:n}}},RESET:c,reset:function(t){return function(){return{type:c(t)}}},ADD:f,add:function(t){return function(n){return{type:f(t),payload:n}}},UPDATE:d,update:function(t){return function(n){return{type:d(t),payload:n}}},REMOVE:s,remove:function(t){return function(n){return{type:s(t),payload:n}}}}),y=function(t){return function(r){return function(e){var u=e;return void 0!==t&&(u=n(e,t)[0]),u[r]}}},p=function(t){return function(n){return function(r){return function(e){return y(n)(r)(e)[t]}}}},v=p("keys"),g=p("array"),h=p("initialized"),m=p("data"),b=Object.freeze({getState:y,getKeys:v,getAsArray:g,getLength:function(t){return function(n){return function(r){return v(t)(n)(r).length}}},isInitialized:h,get:function(t){return function(n){return function(r){return function(e){var u=m(t)(n)(e);return r?Array.isArray(r)?r.map(function(t){return u[t]}):u[r]:u}}}},getBy:function(t){return function(r){return function(e,u){return function(i){var a=g(t)(r)(i);return Array.isArray(u)?a.filter(function(t){return u.includes(n(t,e)[0])}):a.filter(function(t){return u===n(t,e)[0]})}}}}}),A=function(t,n,r){return n in t?Object.defineProperty(t,n,{value:r,enumerable:!0,configurable:!0,writable:!0}):t[n]=r,t},O=Object.assign||function(t){for(var n=1;n<arguments.length;n++){var r=arguments[n];for(var e in r)Object.prototype.hasOwnProperty.call(r,e)&&(t[e]=r[e])}return t},E=function(t){if(Array.isArray(t)){for(var n=0,r=Array(t.length);n<t.length;n++)r[n]=t[n];return r}return Array.from(t)},j={data:{},keys:[],array:[],initialized:!1},k=function(t){return function(n,r){return t.array.find(function(t){return t[n]===r})}},w=function(t){return function(n){return function(){var a=arguments.length>0&&void 0!==arguments[0]?arguments[0]:j,l=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},y=l.type,p=l.payload;switch(y){case o(n):return{data:i(p,t),keys:e(p.map(function(n){return n[t]})),array:p,initialized:!0};case f(n):var v=void 0,g=p[t];return v=k(a)(t,g)?a.array.map(function(n){return n[t]===g?p:n}):[].concat(E(a.array),[p]),O({},a,{data:O({},a.data,A({},g,p)),keys:e([].concat(E(a.keys),[g])),array:v,initialized:!0});case d(n):var h=p[t];return k(a)(t,h)?O({},a,{data:O({},a.data,A({},h,O({},a.data[h],p))),array:a.array.map(function(n){return n[t]===h?O({},n,p):n})}):a;case s(n):return O({},a,{data:r(a.data,[p]),keys:u(a.keys,p),array:a.array?a.array.filter(function(n){return n[t]!==p}):[]});case c(n):return j;default:return a}}}},z={middlewares:[function(t){return function(n){return function(){var r=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return O({},r,{state:w(t)(n)(r.state,r.action)})}}}],actions:["set","add","reset","remove"],selectors:["getState","getKeys","getAsArray","getLength","isInitialized","get","getBy"]},_={data:void 0,initialized:!1},q=function(){return function(t){return function(){var n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:_,r=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},e=r.type,u=r.payload;switch(e){case o(t):return{data:u,initialized:!0};case c(t):return _;default:return n}}}},S={middlewares:[function(t){return function(t){return function(){var n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return O({},n,{state:q()(t)(n.state,n.action)})}}}],actions:["set","reset"],selectors:["getState","isInitialized","get"]},T=Object.freeze({map:z,uniq:S}),x=function(t){return function(n){return function(r){return function(e){var u=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},i=u.type,a={state:e,action:{type:void 0===i?"UNKNOWN":i,payload:u.payload}};return[].concat(E(t.pre||[]),E(t.engine||[]),E(t.post||[])).map(function(t){return t(n)(r)}).forEach(function(t){a=t(a)}),a.state}}}},D=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(n){var r=n.key,e=n.path,u=n.type,i=void 0===u?"map":u,a=n.prefix,o=void 0===a?"":a,c=n.name,f=T[i];return Object.assign.apply(Object,[x(O({},t,{engine:f.middlewares}))(r)(""+o+c),{trampssType:i}].concat(E(Object.keys(l).filter(function(t){return f.actions.includes(t)}).map(function(t){return A({},t,l[t](""+o+c))})),E(Object.keys(b).filter(function(t){return f.selectors.includes(t)}).map(function(t){return A({},t,b[t](e)(c))}))))}},B=function(){throw Error("parameter is not a middleware configuration, nor a factory option object.")},P=function(t){return t.engine||t.pre||t.post},R=function(t){return!!t.name||"string"==typeof t},U=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(n){return null!==n&&void 0!==n||B(),P(n)?function(r){return D(n)(O({},r,t))}:R(n)?"string"==typeof n?D()(O({name:n},t)):D()(O({},n,t)):void B()}},I=U({type:"uniq"}),K=U({type:"map"}),M=U();t.uniq=I,t.map=K,t.default=M,Object.defineProperty(t,"__esModule",{value:!0})});
+!function(t,n){"object"==typeof exports&&"undefined"!=typeof module?n(exports,require("lodash/at"),require("lodash/omit"),require("lodash/uniq"),require("lodash/without"),require("lodash/keyBy")):"function"==typeof define&&define.amd?define(["exports","lodash/at","lodash/omit","lodash/uniq","lodash/without","lodash/keyBy"],n):n(t["trampss-redux-factory"]=t["trampss-redux-factory"]||{},t._at,t._omit,t._uniq,t._without,t._keyBy)}(this,function(t,n,e,r,u,i){"use strict";n=n&&"default"in n?n.default:n,e=e&&"default"in e?e.default:e,r=r&&"default"in r?r.default:r,u=u&&"default"in u?u.default:u,i=i&&"default"in i?i.default:i;var a=function(t){return"@trampss/"+t.toUpperCase()},o=function(t){return a("SET_"+t)},c=function(t){return a("RESET_"+t)},f=function(t){return a("ADD_"+t)},d=function(t){return a("UPDATE_"+t)},s=function(t){return a("REMOVE_"+t)},l=Object.freeze({SET:o,set:function(t){return function(n){return{type:o(t),payload:n}}},RESET:c,reset:function(t){return function(){return{type:c(t)}}},ADD:f,add:function(t){return function(n){return{type:f(t),payload:n}}},UPDATE:d,update:function(t){return function(n){return{type:d(t),payload:n}}},REMOVE:s,remove:function(t){return function(n){return{type:s(t),payload:n}}}}),y=function(t){return function(e){return function(r){var u=r;return void 0!==t&&(u=n(r,t)[0]),u[e]}}},p=function(t){return function(n){return function(e){return function(r){return y(n)(e)(r)[t]}}}},v=p("keys"),g=p("array"),h=p("initialized"),m=p("data"),b=Object.freeze({getState:y,getKeys:v,getAsArray:g,getLength:function(t){return function(n){return function(e){return v(t)(n)(e).length}}},isInitialized:h,get:function(t){return function(n){return function(e){return function(r){var u=m(t)(n)(r);return e?Array.isArray(e)?e.map(function(t){return u[t]}):u[e]:u}}}},getBy:function(t){return function(e){return function(r,u){return function(i){var a=g(t)(e)(i);return Array.isArray(u)?a.filter(function(t){return u.includes(n(t,r)[0])}):a.filter(function(t){return u===n(t,r)[0]})}}}}}),O=function(t,n,e){return n in t?Object.defineProperty(t,n,{value:e,enumerable:!0,configurable:!0,writable:!0}):t[n]=e,t},k=Object.assign||function(t){for(var n=1;n<arguments.length;n++){var e=arguments[n];for(var r in e)Object.prototype.hasOwnProperty.call(e,r)&&(t[r]=e[r])}return t},j=function(t){if(Array.isArray(t)){for(var n=0,e=Array(t.length);n<t.length;n++)e[n]=t[n];return e}return Array.from(t)},A={data:{},keys:[],array:[],initialized:!1},E=function(t){return function(n,e){return t.array.find(function(t){return t[n]===e})}},w=function(t){return function(n){return function(){var a=arguments.length>0&&void 0!==arguments[0]?arguments[0]:A,l=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},y=l.type,p=l.payload;switch(y){case o(n):return{data:i(p,t),keys:r(p.map(function(n){return n[t]})),array:p,initialized:!0};case f(n):var v=void 0,g=p[t];return v=E(a)(t,g)?a.array.map(function(n){return n[t]===g?p:n}):[].concat(j(a.array),[p]),k({},a,{data:k({},a.data,O({},g,p)),keys:r([].concat(j(a.keys),[g])),array:v,initialized:!0});case d(n):var h=p[t];return E(a)(t,h)?k({},a,{data:k({},a.data,O({},h,k({},a.data[h],p))),array:a.array.map(function(n){return n[t]===h?k({},n,p):n})}):a;case s(n):return k({},a,{data:e(a.data,[p]),keys:u(a.keys,p),array:a.array?a.array.filter(function(n){return n[t]!==p}):[]});case c(n):return A;default:return a}}}},z={middlewares:[function(t){return function(n){return function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return k({},e,{state:w(t)(n)(e.state,e.action)})}}}],actions:["set","add","reset","remove"],selectors:["getState","getKeys","getAsArray","getLength","isInitialized","get","getBy"]},_={data:void 0,initialized:!1},q=function(){return function(t){return function(){var n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:_,e=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},r=e.type,u=e.payload;switch(r){case o(t):return{data:u,initialized:!0};case c(t):return _;default:return n}}}},S={middlewares:[function(t){return function(t){return function(){var n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return k({},n,{state:q()(t)(n.state,n.action)})}}}],actions:["set","reset"],selectors:["getState","isInitialized","get"]},T=Object.freeze({keyValue:z,simpleObject:S}),x=function(t){return function(n){return function(e){return function(r){var u=arguments.length>1&&void 0!==arguments[1]?arguments[1]:{},i=u.type,a={state:r,action:{type:void 0===i?"UNKNOWN":i,payload:u.payload}};return[].concat(j(t.pre||[]),j(t.engine||[]),j(t.post||[])).map(function(t){return t(n)(e)}).forEach(function(t){a=t(a)}),a.state}}}},D=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(n){var e=n.key,r=n.path,u=n.type,i=void 0===u?"keyValue":u,a=n.prefix,o=void 0===a?"":a,c=n.name,f=T[i];return Object.assign.apply(Object,[x(k({},t,{engine:f.middlewares}))(e)(""+o+c),{trampssType:i}].concat(j(Object.keys(l).filter(function(t){return f.actions.includes(t)}).map(function(t){return O({},t,l[t](""+o+c))})),j(Object.keys(b).filter(function(t){return f.selectors.includes(t)}).map(function(t){return O({},t,b[t](r)(c))}))))}},V=function(){throw Error("parameter is not a middleware configuration, nor a factory option object.")},B=function(t){return t.engine||t.pre||t.post},P=function(t){return!!t.name||"string"==typeof t},R=function(){var t=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(n){return null!==n&&void 0!==n||V(),B(n)?function(e){return D(n)(k({},e,t))}:P(n)?"string"==typeof n?D()(k({name:n},t)):D()(k({},n,t)):void V()}},U=R({type:"simpleObject"}),I=R({type:"keyValue"}),K=R();t.simpleObject=U,t.keyValue=I,t.default=K,Object.defineProperty(t,"__esModule",{value:!0})});
 //# sourceMappingURL=index.js.map
