+++ src/middlewares/simpleObject.js
